+++ labs/scheduling-at/simulate.sh
@@ -45,12 +45,34 @@ print_line() {
 }
 
 LAB_NAME=$1
+LOG_LEVEL=$2
+
+if [[ ! "$LOG_LEVEL" =~ ^(none|info|basic|debug)$ ]]; then
+    echo "Error: LOG_LEVEL debe ser 'none', 'info', 'basic' o 'debug'."
+    exit 1
+fi
 
 if [ -f "/tmp/$LAB_NAME.log" ]; then
 	sudo rm -rf /tmp/$LAB_NAME.log
 fi
 
-sudo ansible-playbook -i /usr/local/rh134/labs/server.lab  /usr/local/rh134/labs/$LAB_NAME.yml -e "failover_mode='simulate'" >  /tmp/$LAB_NAME.log  2>&1
+# ==============================================================================
+# EJECUCIÓN DE ANSIBLE SEGÚN EL LOG_LEVEL
+# ==============================================================================
+case "$LOG_LEVEL" in
+    none|info)
+        # INFO AND NONE: Silencioso, todo redirigido al log
+        sudo ansible-playbook -i /usr/local/rh134/labs/server.lab /usr/local/rh134/labs/$LAB_NAME.yml -e "failover_mode='simulate'" > /tmp/$LAB_NAME.log 2>&1
+        ;;
+    basic)
+        # BASIC: Se muestra directamente en pantalla (sin redirigir a archivo)
+        sudo ansible-playbook -i /usr/local/rh134/labs/server.lab /usr/local/rh134/labs/$LAB_NAME.yml -e "failover_mode='simulate'"
+        ;;
+    debug)
+        # DEBUG: Modo muy verboso (-vvvv) en pantalla
+        sudo ansible-playbook -i /usr/local/rh134/labs/server.lab /usr/local/rh134/labs/$LAB_NAME.yml -e "failover_mode='simulate'" -vvvv
+        ;;
+esac
 
 echo "Starting lab."
 echo ""
@@ -63,20 +85,20 @@ sleep 2
 print_line "· Generate reports from servera" "SUCCESS"
 echo ""
 
-if [ -f "/tmp/$LAB_NAME.log" ]; then
+if [ "$LOG_LEVEL" = "info" ] && [ -f "/tmp/$LAB_NAME.log" ]; then
 	cat /tmp/$LAB_NAME.log | tail -3
 fi
 
 printf "%-5s \n"
 
-filtered=$(awk -F: '{ NF--; print $0 }' OFS=: /tmp/$LAB_NAME/simulate.log)
+if [ -f "/tmp/$LAB_NAME/simulate.log" ]; then
 
-checked=$(printf '%s\n' "$filtered" | awk -F: '{last=$NF; NF--; print $0 "|" last}' OFS=:)
+	filtered=$(awk -F: '{ NF--; print $0 }' OFS=: /tmp/$LAB_NAME/simulate.log)
 
-while IFS='|' read -r p1 p2; do
-    print_line "· $p1" "$p2"
-    sleep 1
-done <<< "$checked"
+	checked=$(printf '%s\n' "$filtered" | awk -F: '{last=$NF; NF--; print $0 "|" last}' OFS=:)
 
-
-
+	while IFS='|' read -r p1 p2; do
+    		print_line "· $p1" "$p2"
+    		sleep 1
+	done <<< "$checked"
+fi
